[mess] re-enabling dispatch handeler
~ splitting feature and unit tests
